Cambio del .cpp al .h
Cambiar las declaraciones de las clases del archivo piezas.cpp al archivo de piezas.h, que por error se habían ubicado en el archivo incorrecto en uno de los últimos commits.
