--- Piezas.cpp
@@ -1,31 +1,0 @@
-/*
-* En este archivo se crean las clases de cada una de las piezas.
-* Previsiblemente se creará una clase pieza que haciendo uso de hererncia
-* agrupará los aspectos comunes a todas las piezas.
-*/
-
-#include "Posicionamiento.h" //Incluye la clase para el uso del sistema de coordenadas
-
-class Torre {
-
-};
-
-class Caballo {
-
-};
-
-class Alfil {
-
-};
-
-class Dama {
-
-};
-
-class Rey {
-
-};
-
-class Peon {
-
-};
